refactor OCI Auth to its own package
